+++ test/test_alexa_app_schema.js
@@ -17,28 +17,122 @@ describe("Alexa", function() {
     });
 
     describe("#schema", function() {
-      beforeEach(function() {
-        testApp.intent("testIntentTwo", {
-          "slots": {
-            "MyCustomSlotType": "CUSTOMTYPE",
-            "Tubular": "AMAZON.LITERAL",
-            "Radical": "AMAZON.US_STATE",
-          },
+      describe("with a minimum intent", function() {
+        beforeEach(function() {
+          testApp.intent("AMAZON.PauseIntent");
         });
 
-        testApp.intent("testIntent", {
-          "slots": {
-            "AirportCode": "FAACODES",
-            "Awesome": "AMAZON.DATE",
-            "Tubular": "AMAZON.LITERAL"
-          },
+        it("contains no slots", function() {
+          var subject = JSON.parse(testApp.schema());
+          expect(subject).to.eql({
+            "intents": [{
+              "intent": "AMAZON.PauseIntent"
+            }]
+          });
         });
       });
 
-      it("generates the expected schema", function() {
-        var expected = mockHelper.load("expected_intent_schema.json");
-        var subject = JSON.parse(testApp.schema());
-        expect(subject).to.eql(expected);
+      describe("with empty slots", function() {
+        beforeEach(function() {
+          testApp.intent("AMAZON.PauseIntent", {
+            "slots": {}
+          });
+        });
+
+        it("contains no slots", function() {
+          var subject = JSON.parse(testApp.schema());
+          expect(subject).to.eql({
+            "intents": [{
+              "intent": "AMAZON.PauseIntent"
+            }]
+          });
+        });
+      });
+
+      describe("with a slot", function() {
+        beforeEach(function() {
+          testApp.intent("testIntent", {
+            "slots": {
+              "MyCustomSlotType": "CUSTOMTYPE",
+              "Tubular": "AMAZON.LITERAL",
+              "Radical": "AMAZON.US_STATE",
+            },
+          });
+        });
+
+        it("includes slots", function() {
+          var subject = JSON.parse(testApp.schema());
+          expect(subject).to.eql({
+            "intents": [{
+              "intent": "testIntent",
+              "slots": [{
+                "name": "MyCustomSlotType",
+                "type": "CUSTOMTYPE"
+              }, {
+                "name": "Tubular",
+                "type": "AMAZON.LITERAL"
+              }, {
+                "name": "Radical",
+                "type": "AMAZON.US_STATE"
+              }]
+            }]
+          });
+        });
+      });
+
+      describe("with multiple intents", function() {
+        beforeEach(function() {
+          testApp.intent("AMAZON.PauseIntent");
+
+          testApp.intent("testIntentTwo", {
+            "slots": {
+              "MyCustomSlotType": "CUSTOMTYPE",
+              "Tubular": "AMAZON.LITERAL",
+              "Radical": "AMAZON.US_STATE",
+            },
+          });
+
+          testApp.intent("testIntent", {
+            "slots": {
+              "AirportCode": "FAACODES",
+              "Awesome": "AMAZON.DATE",
+              "Tubular": "AMAZON.LITERAL"
+            },
+          });
+        });
+
+        it("generates the expected schema", function() {
+          var subject = JSON.parse(testApp.schema());
+          expect(subject).to.eql({
+            "intents": [{
+              "intent": "AMAZON.PauseIntent",
+            }, {
+              "intent": "testIntentTwo",
+              "slots": [{
+                "name": "MyCustomSlotType",
+                "type": "CUSTOMTYPE"
+              }, {
+                "name": "Tubular",
+                "type": "AMAZON.LITERAL"
+              }, {
+                "name": "Radical",
+                "type": "AMAZON.US_STATE"
+              }]
+            }, {
+              "intent": "testIntent",
+              "slots": [{
+                "name": "AirportCode",
+                "type": "FAACODES"
+              }, {
+                "name": "Awesome",
+                "type": "AMAZON.DATE"
+              }, {
+                "name": "Tubular",
+                "type": "AMAZON.LITERAL"
+              }]
+            }]
+          });
+        });
       });
     });
   });